Moved {cell_width, padding} from HbProfile to collider::Collider::new().
cell_width() and padding() were functions that didn't accept any arguments.  As
a result, they could only return hardcoded values (or lookup values in a global
dictionary, which would be awkward).  This commit removes both cell_width() and
padding() functions from the HbProfile trait, and adds them as arguments to
collider::Collider::new().  This allows you to create/destroy Collider instances
on the fly, which can be useful in certain types of simulators.